Successfully build first FreeRTOS Task
