+++ .gitignore
@@ -1,2 +1,3 @@
 **/cmake-*/
 build/
+**.bak
