+++ .vscode/settings.json
@@ -1,4 +1,12 @@
 {
     "cmake.sourceDirectory": "${workspaceFolder}/dert",
-    "C_Cpp.default.includePath": [ "${workspaceFolder}/dert/include"]
+    "C_Cpp.default.includePath": [
+        "${workspaceFolder}/dert/include"
+    ],
+    "files.associations": {
+        "stdio.h": "c",
+        "gpio.h": "c",
+        "i2c.h": "c",
+        "stdlib.h": "c"
+    }
 }
