+++ dert/include/dert_sense_light_task.h
@@ -1,0 +1,7 @@
+/**
+ * Copyright (c) 2022 Kris Keillor
+ *
+ * SPDX-License-Identifier: BSD-3-Clause
+ */
+
+void vDertSenseLight(void *pvParameters);
