+++ dert/src/freertos/CMakeLists.txt
@@ -1,18 +1,21 @@
 if (TARGET tinyusb_device)
-    add_executable(dert_freertos
-            dert_freertos.c
+    add_executable(main
+            main.c
             )
 
     # pull in common dependencies
-    target_link_libraries(dert_freertos pico_stdlib pico_multicore hardware_i2c FreeRTOS-Kernel FreeRTOS-Kernel-Heap4)
+    target_link_libraries(main pico_stdlib pico_multicore hardware_i2c FreeRTOS-Kernel FreeRTOS-Kernel-Heap4)
+
+    # pull in sources
+    target_sources(main PRIVATE dert_sense_air_task.c)
 
     # enable usb output, disable uart output
-    pico_enable_stdio_usb(dert_freertos 1)
-    pico_enable_stdio_uart(dert_freertos 0)
+    pico_enable_stdio_usb(main 1)
+    pico_enable_stdio_uart(main 0)
 
     # create map/bin/hex/uf2 file etc.
-    pico_add_extra_outputs(dert_freertos)
+    pico_add_extra_outputs(main)
 
 elseif(PICO_ON_DEVICE)
-    message(WARNING "not building dert_freertos because TinyUSB submodule is not initialized in the SDK")
+    message(WARNING "not building main because TinyUSB submodule is not initialized in the SDK")
 endif()
