+++ dert/src/freertos/dert_freertos.c
@@ -13,55 +13,6 @@
 
 #include "dert_2040.h"
 
-int main() {
-    stdio_init_all();
-    int lifespan = 0;
-    int state = Startup;
-
-    // Light sensor fields
-    uint8_t bh1750_dat_tx;
-    uint8_t bh1750_dat_rx[2];
-    int bh1750_dat_err;
-
-    // Air sensor fields
-    uint8_t sht30_dat_tx;
-    uint8_t sht30_dat_rx[2];
-    int8_t sht30_dat;
-
-    // Soil sensor fields
-    uint8_t moisture_sns_dat_tx;
-    uint8_t moisture_sns_dat_rx[2];
-    int moisture_sns_dat_err;
-
-    // Wait for PuTTY to connect (USB-only)
-    sleep_ms(5000);
-
-    printf("    DERT state: Starting up!\n");
-    gpio_init(GPIO_I2C_SDA);
-    gpio_init(GPIO_I2C_SCL);
-    i2c_init(i2c_default, 100*1000);
-    gpio_set_function(GPIO_I2C_SDA, GPIO_FUNC_I2C);
-    gpio_set_function(GPIO_I2C_SCL, GPIO_FUNC_I2C);
-    bi_decl(bi_2pins_with_func(GPIO_I2C_SDA, GPIO_I2C_SCL, GPIO_FUNC_I2C));
-
-    gpio_init(GPIO_LED0);
-    gpio_init(GPIO_LED1);
-
-    gpio_init(GPIO_HVR);
-    gpio_init(GPIO_LVR1);
-    gpio_init(GPIO_LVR2);
-    gpio_set_dir(GPIO_HVR, GPIO_OUT);
-    gpio_set_dir(GPIO_LVR1, GPIO_OUT);
-    gpio_set_dir(GPIO_LVR2, GPIO_OUT);
-
-    gpio_init(GPIO_DB1);
-    gpio_init(GPIO_DB2);
-    gpio_init(GPIO_DB3);
-
-    // Initialize FreeRTOS application
-    vTaskStartScheduler();
-}
-
 // Exclude reserved addresses, of the form 000 0xxx or 
 // 111 1xxx, from the search.
 bool i2c_res_addr(uint8_t addr) {
