+++ dert/src/freertos/dert_report_data_task.c
@@ -1,3 +1,18 @@
+/**
+ * Copyright (c) 2022 Kris Keillor
+ *
+ * SPDX-License-Identifier: BSD-3-Clause
+ */
+
+#include <stdio.h>
+
+#include "pico/stdlib.h"
+#include "pico/binary_info.h"
+#include "hardware/gpio.h"
+#include "hardware/i2c.h"
+
+#include "dert_2040.h"
+
 static void vDertReportData(void *pvParameters) {
     for ( ;; ) {
         printf("    DERT state: Reporting data!\n");
